refactor(term): add custom table widget to remove duplicate code
There were duplicate sections in the process of generating the `ratatui` Table widget that was used for rendering the list of GH notifications, entries, and feeds.
To maintain the same UX, it was necessary to modify each of these individually when changing the table rendering process.
Therefore, We extracted a custom table widget and centralized the configuration process.