chore: reorganize release tasks
We used to generate CHANGELOG with git-cliff and execute cargo-release separately.
In this case, two commits were created during the release process, one for changelog update
and another for the version bump.

To address this, we decided to generate the changelog in the `pre-release-hook` of cargo-release
which allows us to combine theses commits.

Since the `pre-release-hook` set the crate as the working directory, we relocated the just tasks
for each crate accordingly.